services/user/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-consul==1.1.0
requests==2.28.2
python-jose[cryptography]==3.3.0
pydantic==1.10.7
